Folder contains the objects and main program for a basic ecosystem simulation
The objects relate to a Plant Class, Herbivore Class, and Carnivore Class.  WildLand utilizes these objects to simulate an ecosystem in which there is a relation of organism and food source.  The user does not instantiate the objects.  Rather, the user defines the parameters of the simulation and the objects are dynamically created based on those parameters.  Images of the output and a class diagram are also included.